VIITE-82 use generic mapper trait and migrate code to trait
